+++ idea/testData/spring/core/generate/beanDependenciesByXml/setter.kt
@@ -1,0 +1,9 @@
+// ACTION_CLASS: org.jetbrains.kotlin.idea.spring.generate.GenerateKotlinSpringBeanDependencyAction$Setter
+// CONFIG_FILE: setter-config.xml
+// CHOOSE_BEAN: barBean
+package a
+
+open class FooBean {<caret>
+}
+
+open class BarBean